Improve auth page UI.
