+++ client/src/App.css
@@ -2,6 +2,7 @@
   --dark-gray-navbar: #202225;
   --dark-gray-navbar-icon: #36393f;
   --light-gray-icon: #36393f;
+  --light-blue-azure: #007fff;
   --green-navbar-icon: #3ba55d;
   --dark-gray-sidebar: #2f3136;
   --dark-gray-sidebar-hover: #393c43;
@@ -26,7 +27,6 @@ body {
 }
 /*  */
 .auth__form-container_fields {
-  flex: 2;
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -53,6 +53,7 @@ body {
   justify-content: flex-start;
 
   padding: 2rem;
+  width: 30%;
   box-shadow: 0px 1px 5px rgb(0 0 0 / 10%);
   border-radius: 5px;
   transition: 0.8s ease;
@@ -90,7 +91,7 @@ body {
   font-size: 14px;
   outline: none;
   transition: all 150ms ease-in-out 0s;
-  width: 85%;
+  width: 100%;
   background: #fff;
 }
 /*  */
@@ -119,15 +120,13 @@ body {
 }
 /*  */
 .auth__form-container_fields-content_button {
-  margin-top: 2rem;
   display: flex;
   justify-content: flex-start;
 }
 /*  */
 .auth__form-container_fields-content_button button {
   border-radius: 4px;
-  background: #005fff;
-  border: 1px solid #005fff;
+  background: var(--light-blue-azure);
   color: #fff;
   font-family: Arial, Helvetica, sans-serif;
   font-weight: 500;
@@ -146,7 +145,7 @@ body {
   justify-content: flex-start;
   align-items: center;
 
-  margin-top: 0.2rem;
+  margin-top: 0.45rem;
 }
 /*  */
 .auth__form-container_fields-account p {
