+++ client/src/components/Auth.jsx
@@ -4,6 +4,8 @@ import Cookies from "universal-cookie";
 import axios from "axios";
 
 import signinImage from "../assets/signup.jpg";
+import Navbar from "./Navbar";
+import Footer from "./Footer";
 
 const cookies = new Cookies();
 
@@ -17,7 +19,7 @@ const initialState = {
 
 const API_ROUTE = import.meta.env.VITE_API_ROUTE;
 
-const Auth = () => {
+const Auth = ({ userProfile, setUserProfile }) => {
   const authToken = cookies.get("jwtToken");
   const navigate = useNavigate();
   const location = useLocation();
@@ -67,41 +69,40 @@ const Auth = () => {
   if (authToken) navigate("/workspace");
 
   return (
-    <div className="auth__form-container">
-      <div className="auth__form-container_fields">
-        <div className="auth__form-container_fields-content">
-          <p>{isSignup ? "Sign Up" : "Sign In"}</p>
-          <form onSubmit={handleSubmit}>
-            <div className="auth__form-container_fields-content_input">
-              <label htmlFor="username">Username</label>
-              <input type="text" name="username" placeholder="Username" onChange={handleChange} required />
-            </div>
-            {isSignup && (
-              <div className="auth__form-container_fields-content_input">
-                <label htmlFor="email">Email</label>
-                <input type="email" name="email" placeholder="abc@abc.com" onChange={handleChange} required />
-              </div>
-            )}
-            <div className="auth__form-container_fields-content_input">
-              <label htmlFor="password">Password</label>
-              <input type="password" name="password" placeholder="Password" onChange={handleChange} required />
-            </div>
-            <div className="auth__form-container_fields-content_button">
-              <button> {isSignup ? "Sign Up" : "Sign In"}</button>
-            </div>
-          </form>
-          <div className="auth__form-container_fields-account">
-            <p>
-              {isSignup ? "Already have an account?" : "Don't have an account?"}
-              <span onClick={switchMode}>{isSignup ? "Sign In" : "Sign Up"}</span>
-            </p>
+    <main className="w-full min-h-screen flex flex-col justify-between items-center bg-dark-color-blue-background">
+      <Navbar userProfile={userProfile} setUserProfile={setUserProfile} />
+
+      <div className="auth__form-container_fields-content">
+        <p>{isSignup ? "Sign Up" : "Sign In"}</p>
+        <form onSubmit={handleSubmit}>
+          <div className="auth__form-container_fields-content_input">
+            <label htmlFor="username">Username</label>
+            <input type="text" name="username" placeholder="Username" onChange={handleChange} required />
           </div>
+          {isSignup && (
+            <div className="auth__form-container_fields-content_input">
+              <label htmlFor="email">Email</label>
+              <input type="email" name="email" placeholder="abc@abc.com" onChange={handleChange} required />
+            </div>
+          )}
+          <div className="auth__form-container_fields-content_input">
+            <label htmlFor="password">Password</label>
+            <input type="password" name="password" placeholder="Password" onChange={handleChange} required />
+          </div>
+          <div className="auth__form-container_fields-content_button">
+            <button> {isSignup ? "Sign Up" : "Sign In"}</button>
+          </div>
+        </form>
+        <div className="auth__form-container_fields-account">
+          <p>
+            {isSignup ? "Already have an account? " : "Don't have an account? "}
+            <span onClick={switchMode}>{isSignup ? "Sign In" : "Sign Up"}</span>
+          </p>
         </div>
       </div>
-      <div className="auth__form-container_image">
-        <img src={signinImage} alt="sign in" />
-      </div>
-    </div>
+
+      <Footer />
+    </main>
   );
 };
