+++ client/src/components/Footer.jsx
@@ -1,0 +1,9 @@
+const Footer = () => {
+  return (
+    <footer className="w-full text-center mb-3 text-light-color-gray text-sm">
+      &copy; Appworks School Batch #20 Backend Class. Hosted by Shao-Ting Fong.
+    </footer>
+  );
+};
+
+export default Footer;
